Initial commit with basic commands and structure
Goal - support optional initialisation with eval "$(pfm init -)" but
still have everything worh without this as long as PFM_HOME is defined.

:; is used to mark commands which should only be run in uninitialised
shells - in general this should just delegate to the function defined in
the file as the last line but certain env modifying commands like init
will need it for other use cases.